refactor: 패키지 구조 변경 catalog 패키지 제거
