+++ src/main/java/com/feellog/backend/domain/category/entity/CategoryGroup.java
@@ -1,4 +1,4 @@
-package com.feellog.backend.domain.catalog.situationtag.entity;
+package com.feellog.backend.domain.category.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -10,12 +10,12 @@ import java.time.LocalDateTime;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "situation_tag")
-public class SituationTag {
+@Table(name = "category_group")
+public class CategoryGroup {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "situation_tag_id")
+    @Column(name = "category_group_id")
     private Long id;
 
     @Column(nullable = false, unique = true, length = 100)
