+++ src/main/java/com/feellog/backend/domain/paymentmethod/repository/PaymentMethodRepository.java
@@ -1,0 +1,12 @@
+package com.feellog.backend.domain.paymentmethod.repository;
+
+import com.feellog.backend.domain.paymentmethod.entity.PaymentMethod;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PaymentMethodRepository extends JpaRepository<PaymentMethod, Long> {
+
+    Optional<PaymentMethod> findByName(String name);
+
+}
