+++ src/main/java/com/feellog/backend/domain/situationtag/entity/SituationTag.java
@@ -1,4 +1,4 @@
-package com.feellog.backend.domain.catalog.paymentmethod.entity;
+package com.feellog.backend.domain.situationtag.entity;
 
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -10,15 +10,15 @@ import java.time.LocalDateTime;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "payment_method")
-public class PaymentMethod {
+@Table(name = "situation_tag")
+public class SituationTag {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "payment_method_id")
+    @Column(name = "situation_tag_id")
     private Long id;
 
-    @Column(nullable = false, unique = true, length = 50)
+    @Column(nullable = false, unique = true, length = 100)
     private String name;
 
     @Column(nullable = false)
